requirements.txt
fastapi~=0.111.0
uvicorn
pydantic~=2.7.1
sqlmodel
SQLAlchemy
python-jose[cryptography]
passlib[bcrypt]~=1.7.4